feat(core): rename renderer
