+++ packages/core/src/app-switcher/logic-renderer/interface.ts
