+++ packages/core/src/app-switcher/logic-renderer/service.ts
