+++ packages/core/src/hooks/interface.ts
@@ -6,12 +6,12 @@ import { createServiceSymbol } from '../utils';
 export const IHooksKey = createServiceSymbol('IHooks');
 
 export interface IHooks {
-  /** 在执行加载应用之前的勾子 */
-  beforeLoadApps: AsyncSeriesHook<ILogicLoaderHookContext>;
+  /** 在执行逻辑加载应用之前的勾子 */
+  beforeLogicLoad: AsyncSeriesHook<ILogicLoaderHookContext>;
 
-  /** 执行加载应用的勾子 */
-  loadApps: AsyncSeriesHook<ILogicLoaderHookContext>;
+  /** 执行逻辑加载应用的勾子 */
+  logicLoad: AsyncSeriesHook<ILogicLoaderHookContext>;
 
-  /** 在执行加载应用之后的勾子 */
-  afterLoadApps: AsyncSeriesHook<ILogicLoaderHookContext>;
+  /** 在执行逻辑加载应用之后的勾子 */
+  afterLogicLoad: AsyncSeriesHook<ILogicLoaderHookContext>;
 }
