+++ packages/core/src/hooks/service.ts
@@ -8,9 +8,9 @@ export * from './interface';
 
 @provide(IHooksKey)
 export class Hooks implements IHooks {
-  public beforeLoadApps = new AsyncSeriesHook<ILogicLoaderHookContext>();
+  public beforeLogicLoad = new AsyncSeriesHook<ILogicLoaderHookContext>();
 
-  public loadApps = new AsyncSeriesHook<ILogicLoaderHookContext>();
+  public logicLoad = new AsyncSeriesHook<ILogicLoaderHookContext>();
 
-  public afterLoadApps = new AsyncSeriesHook<ILogicLoaderHookContext>();
+  public afterLogicLoad = new AsyncSeriesHook<ILogicLoaderHookContext>();
 }
